Update patch set 1
Patch Set 1: Code-Review-1

(35 comments)

The three include scripts that are being pulled in are all very similar. It would make sense to merged them down as much as possible and only have a snippet that was custom to each of them that was included along with the rest of the script. This isn't something that needs to happen now, but would be a good idea in the future.

Patch-set: 1
Label: Code-Review=-1