[CIMAN-6] 'mso' is now called 'so'
The 'mso' repositories in gerrit were deprecated (marked read-only)
and new 'so' repositories were created. The jenkins jobs should build
the code from the 'so' repositories, not the 'mso' repositories.

NOTE TO REVIEWERS: I don't see any direct references to the yaml files
I've renamed.  Please confirm that this change is correct.

Signed-off-by: Rob Daugherty <rd472p@att.com>
Change-Id: Ibefd42d289c76f927e7541ec0886235f02187a42