Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

> This adjusts Jenkins to build and deploy the portal sdk jars from
 > the new repo portal/sdk using file name portal-sdk.yaml.
 > 
 > Does a new Jenkins job just spring into existence or is there a
 > control file I must also modify?  How do we turn off the old one? 
 > Should I remove the old YAML file entirely?

Upon merge of JJB any jobs that do not exist in Jenkins will automatically be created. If a job is removed it will be deleted. If a job is modified then it will be updated.

 > Also some questions about the new config file:
 > 
 > 1. I modified it to have a settings entry but that does not exist
 > yet - how do I create a settings file?
 > 
 > mvn-settings: 'portal-sdk-settings'

The ${project}-settings files are maintained by LF Release Engineering as they are the publishing credentials into Nexus for the given project.

 > 2. There have been no releases yet of this one, so I removed these
 > entries
 > 
 > - 'release-1.0.0':
 > branch: 'release-1.0.0'
 > - 'release-1.1.0':
 > branch: 'release-1.1.0'

That was the correct thing to do.

Patch-set: 1
Label: Code-Review=-1