Update patch set 3
Patch Set 3:

Thank's Gary, sorry if I am asking many questions, but this change is confusing to me. Is this job for policy? or for integration?
If this is for the integration repo, wouldn't this be part of the integration yaml file?

Patch-set: 3
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0