Update patch set 1
Patch Set 1:

Just because it's failing to build is no excuse for abandoning a change. It's always better to just work the issue until it passes than create new changes in Gerrit. Don't be ashamed of failures showing up, they happen!

The only time I personally believe a change should be abandoned is if it's truly a bad change that won't get merged or is in direct conflict with a better change / one that already fixes the issue that's trying to be addressed.

Patch-set: 1
Reviewer: Gerrit User 1 <1@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0