Update patch set 2
Patch Set 2:

Line 11 "yum install"?  That is only run for CentOS which wouldn't apply in this case.  If the apt-get version works, then we should replace the entire contents of the ubuntu_systems() function with "apt-get -y install redis-server".  I think the apt-get installation would setup redis-server to start on boot.

Patch-set: 2