Add support for repo.git submodule directories
Some ONAP projects use a node in the repository path to represent both
hierarchy for lower level repositories and an actual respository.
To support this in the doc project submodules directory structure we will
append .git to every repository. For example, here are three directories
two of which represent repositories appc.git appc/deployment.git.

The doc-master-verify-rtd template has been modified to work with repo.git
naming when it exist and remain compatible repo naming until .git is added
to all repositories

Change-Id: I0b97fe71e129afaa400bef3385c4cc8d2d525c49
Issue-ID: DOC-33
Signed-off-by: Rich Bennett <rb2745@att.com>