Update patch set 2
Patch Set 2:

> Ok, in that case, can you guys help set this up?  I don't really
 > know what system-scripts, resource type, system_matrix, etc. mean
 > that you were referring to.

jenkins-scripts == the jenkins-scripts directory in this repository. If you open the system_type.sh script you'll see an for loop using an IFS of ',' that's the system matrix that was being referred to.

The docker bits will need to be put into a docker.sh file which will be a new file.

system_type.sh is a switching mechanism for determining a portion of a hostname that should map to a configuration script. This is how we map the provisioning / template to any specialized startup scripts and is why our minions all encode their packer template information into their name.

Patch-set: 2