Fix TCA release job
Daily release build only deposits SNAPSHOT artifacts for TCA.
Lusheng investigated and found that TCA jjb is not calling the standard
java daily release job template.  Instead it uses a template without
setting the release versions, and tries to set the release version by
itself.  However that is not done correctly.  So in TCA daily release
build the release version is not set correctly, and as a result no
release artifact is built, only the SNAPSHOT versioning.

To fix this, we change to use
'{project-name}-{stream}-release-version-java-daily' job.

Change-Id: I59c4962073108a43b0755049058f04b106ca5e6e
Signed-off-by: Nekrassov, Alexei (an4828) <nekrassov@att.com>
Issue-ID: DCAEGEN2-123
Signed-off-by: Nekrassov, Alexei (an4828) <nekrassov@att.com>