Update patch set 1
Patch Set 1:

I understand that. But i need to NOT run sonar jobs for the amsterdam branch. Otherwise they just run over each other depending on the branch. We won't be fixing sonar issues for Amsterdam as i understand it. If there is another solution I'd like to know.

Patch-set: 1
Label: Code-Review=0