Update patch set 1
Patch Set 1:

Jessica, how are we supposed to handle this then? Are we supposed to put only the master stream in the master branch ci-management and in amsterdam branch change the stream to use that branch.

Patch-set: 1
Reviewer: Gerrit User 75 <75@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0