Update patch set 2
Patch Set 2:

> Jobs can be remotely triggered via REST API with either a user's
 > credentials, or an auth-token specific to each job.  With the
 > latter, no user credential is necessary.
 > 
 > For these jobs in particular, since all the job is doing is
 > retrieving some log files, I didn't think a strong authentication
 > scheme was strictly necessary.  Using a job token is much more
 > convenient since this eliminates the long latency in supporting
 > user-based authentication; the onap-ci-windriver user still does
 > not have the right permissions despite multiple requests to
 > helpdesk.
 > 
 > So, we have a few options:
 > 1. continue with my patch as is with the auth-tokens
 > world-readable,
 > 2. use auth-tokens but hide the credentials somehow, or
 > 3. go back to user-based authentication but we need much faster
 > turnaround on getting the account permissions set up correctly.

I'm not fond of having any sort of auth credentials hanging out in public. It's just bad practice even if it's for such basic jobs as it sets a bad precedent.

I suggest going back to option 3 and as such I've taken the time to configure the account permission for onap-ci-windriver in both the production and sandbox Jenkins system. We were missing a security group for it and having it defined in Jenkins. Any third-party labs going forward will need to be added to that group at the same time as we grant streaming rights.

Patch-set: 2