Update patch set 2
Patch Set 2:

> I think this change will overwrite the scripts step entirely and
 > replace it by just this tox script.
 > 
 > Did you tested this on the sandbox?

Not yet, I need run UT and coverage during sonar job, I added the exec-maven-plugin in pom.xml and found that all job run UT and coverage, so I moved the sh to jjb directory to see if this can work. If this will overwrite the scripts step, can you tell me how to fix this problem?

Patch-set: 2