Update patch set 1
Patch Set 1:

> is this change meant for music repo or music/distributed-kv-store
 > repo?
 > 
 > https://gerrit.onap.org/r/#/admin/projects/?filter=music

So this is for the music/distributed-kv-store project. Because the functionality for the KV store is under MUSIC in integration repo [1], The project name 'music' is used with the functionality 'music-distributed-kv-store-test-plan' which is the plan present in the integration repo.  

[1] - https://git.onap.org/integration/tree/test/csit/plans/music

Patch-set: 1
Reviewer: Gerrit User 991 <991@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0