Update patch set 5
Patch Set 5:

So, I did some tests in the sandbox and noticed the jobs that you are using have a little bug that I need to fix first. These jobs basically should be taking the "files" variable you are using in line 19. This variable, we need to set it to "files: 'pom.xml'" If you want these jobs to be triggered by changes in pom.xml only. The issue is that this job is hardcoded to pick all file by default. 

Let me do this correction in the templates first and link you to that change. Then after, I can correct yours.

Give me 2 mins..

Patch-set: 5