Update patch set 9
Patch Set 9:

Oh... that one.. 
So that one is part of the post build actions. This command is just for deploying the logs of the jenkins job itself and this command is just standard for all jobs. 

You need to look into the build steps rather. 
The verify job runs the install with the profile action 
The merge job is the one that runs the deploy with the profile action too. 

This logs post build action runs after the actual build (with your profile) succeeds or fails.

Patch-set: 9