Add extra release job for drools-pdp
There is a cyclical dependency between 2 repos:
policy/drools-pdp and policy/drools-applications but
only in the docker build. I need a release job that creates
the drools-pdp artifacts - which I am adding here.
Then I can release the drools-applications and follow
up with building the released docker image for the
drools-pdp.

Issue-ID: POLICY-708
Change-Id: I3e50a3d8b86ab8ba4cb713082e9cef7cc5d6a04e
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>