Update patch set 2
Patch Set 2:

The issue is because there is a gerrit-maven-clm defined for release-2.5.3 in the file music-jar.yaml. 

We can do 1 thing, we can remove that job from music-jar.yaml, and just have it defined in music.yaml. 
But, you need to confirm with the team which is the pom file that you are going to use for the CLM scans, since we have "mvn-params: '-f jar/pom.xml'" in music-jar.yaml but in music.yaml we don't have a specific pom to use for CLM.

Patch-set: 2
Label: Code-Review=0