Move Dmaven.test.skip param (a-e)
Rather than hardcoding this parameter as part of the
maven-docker-push-daily macro, we are passing it as a
"mvn-params" parameter from the project's JJB template.

Notice that this change IS NOT adding "Dmaven.test.skip=true"
but rather moving it away from the macro into the project's
template to handle it individually per project.

Please refer to commit 262e8694ebce04450bc29b4805088a354b8aaccb

Change-Id: I0f2e0d681734939f4e9b7d07faca33cc4a0da2b9
Issue-ID: CIMAN-159
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>