Update patch set 2
Patch Set 2:

> (1 comment)

"Denes,

Can you please try pushing this into 10003 rather than 10004?

I know this is confusing, but chatting with Andy I believe there are old
reasons why 10004 is not being used as a staging repository at all.
Seems like ONAP decided to continue using 10003 for both staging and snapshots in
Nexus3.

Please try that and let's see if that allows you to push.

thanks!"

I tried to add in an other parameter that switches between the build mode and is independent from the docker registry.

Patch-set: 2