Update patch set 2
Patch Set 2:

Yea,, that suggestion was made by me. 
We should pass this variable in your project's YAML not on the job 
template. 

In your project's yaml, you will need to set the varaible 
DOCKERREGISTRY to 10003

I asked Jeremy to give us a hand too since we are little bit overloaded.

Patch-set: 2