Update patch set 1
Patch Set 1:

We have separated the build process per company in the vfc/svnfm. Now each folder in the svnfm project has a separate set of Jenkins jobs. The problem with the current setup is that the Gerri trigger "please merge" triggers a lot more job than it supposed to. The trigger for the per company job should only be triggered it some files are modified in the directory for the company. So all jobs in jjb/vfc/vfc-nfvo-driver-vnfm-svnfm-nokia.yaml file should only be triggered if the change is happening in the nokiav2 directory. This is why I added in the files section a nokiav2/** pattern.

Patch-set: 1