Add template to build release docker image
The release docker image is built from artifacts in the release repo.
This job is only triggered manually by Linux foundation, when a release
needs to be built. The release builder still publishes to the staging repo.
The images is planed to be moved from staging repo to the release repo
manually.

Change-Id: I943494d6826aa8f4e9186faa8c82377aa8c5b09f
Issue-ID: VFC-728
Signed-off-by: Denes Nemeth <denes.nemeth@nokia.com>