Update patch set 7
Patch Set 7: Code-Review-1

Let's hold this change for now. We are just in the process of moving into simpler docker templates after the presentation that Adolfo gave in the Casablanca meetings. We need to move first into a simpler workflow before we can start adopting the cascade builds

Patch-set: 7
Label: Code-Review=-1