Update patch set 7
Patch Set 7:

Dear Jess

I would like to merge this. I have been present on the Docker improvement presentation. I do not think that it will help in our case. The docker images build for Nokia for Beijing release is still inconsistent (and not working) due to the bad build procedure of not packing the right software in the image. Our Docker files are already very simple. If Adolfo implements the improvements than this commit does not affect him in any way, but since there is no commitment "when" and "what" exactly will be implemented I do not really want to wait for this. I would like to create a new release of the Nokia driver with this procedure.

Thanks Denes

Patch-set: 7