Update patch set 1
Patch Set 1: Code-Review-1

Can we split this change. 

We need to isolate the changes to the global files from the project specific yamls. 

Why do we need to duplicate the already existing jobs in 	jjb/global-templates-java.yaml?
Can we use the ones we have already? I am working into reducing the amount of templates we have.

Patch-set: 1
Label: Code-Review=-1