Update patch set 1
Patch Set 1:

This is because many jobs use the original template, and we're trying to gradually migrate some of those jobs to a new template without changing the job names.  Also, not all of these jobs are eligible to migrate to the new template, so we might need to live with both for a while.

Patch-set: 1