Update patch set 2
Patch Set 2: -Code-Review

We need a docker job which doesn't snip the snapshot tag from pom.xml. The requirement is to read the pom.xml and if it contains a SNAPSHOT keyword in the version the docker build script will build images using 'snapshots' maven artifacts and pushes it to 'snapshots' docker repository. If pom.xml version is with out snapshots (for eg: 1.2.0) it will build from 'releases' maven artifact and pushes to 'releases' docker repository. How can I achieve this using a single job?

Patch-set: 2