Update SDC sonar jjb
Removed old style Jave memory definitions.

Add sonar branch name property to be passed to the maven on execution.
Add the use of the branch jjb variable, to allow differentiation of what branch the sonar run on.
Add use of the LF npm proxy, populate the reposetory with the LF repo when runing in LF jenkins.

Change-Id: Ia996bb8e925461c3fa5af6c1ff40a0371faa9123
Issue-ID: SDC-1632
Signed-off-by: Michael Lando <ml636r@att.com>