Expand rendered image name to include sec and ms
Expand rendered image name to include seconds and milliseconds.
This reduces the probability of image name collisions on the cloud
from 1/60 to 1/60000.

The initial change Ic1853ba7566def397b8bce389e2779fe3020c7ce was
reverted, since the change had issues bringing up an instance on the
OpenStack cloud provider. This was because the default `instance_name`
value is taken as the `image_name`, when the instance_name is unset,
the instance_name does not allow FQDN format hostname ending with
a period+number.

The workaround is to use separate `instance_name` parameter in your
builder which does not end with period+number.

Fixed upstream in common-packer [1.].

[1.] https://gerrit.linuxfoundation.org/infra/#/c/10986/

JIRA: https://jira.linuxfoundation.org/browse/RELENG-971

Issue-ID: CIMAN-168
Change-Id: I451dd8956fbd2b97a47b5064d9802108f4fc3115
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>