Update patch set 1
Patch Set 1:

Yes,, let me fix this job.. 

Adding this job into the repos is a long process since not only we need to add it but to make sure the INFO.yaml is clear of any errors. It is a lot of time consuming but I have in my tasks to add this job for all repos. I believe VNFSDK is one of the few teams that have their info jobs working and clear of errors.. Let me fix yours..

Patch-set: 1