MULTICLOUD-Refactor CLM project block
The clm job portion needs to be on its own
project section similar to the Sonar portion.

The reason behind this, is becuase we only want
this job to be running for master and no other
branches. Also, this is in preparation for the
work that is to come from global-jjb on application
names in CLM.

Change-Id: I687796265ae87c0698416c8649efdfd491779107
Issue-ID: CIMAN-213
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>