Update patch set 1
Patch Set 1: Code-Review-1

We do not need to create a copy of {project-name}-{stream}-docker-java-shell-daily. 

Basically, if we want the name of this job to reflect "arm64" on it, we can do it by adding a new project block in the project's yaml file. 

For example, if we need to add aaf-authz project for this job, then we need to create a new block in jjb/aaf/aaf-authz.yaml that will look like this:

- project:
    name: aaf-authz-arm64
    project-name: 'aaf-authz-arm64'
    jobs:
      - '{project-name}-{stream}-docker-java-shell-daily':
          mvn-goals: 'clean install'
          settings: '{mvn-settings}'
          script:
            !include-raw-escape: 'build-aaf-docker-image.sh'
          build-node: 'ubuntu1604-docker-8c-8g'
    project: 'aaf/authz'
    stream:
      - 'master':
          branch: 'master'
    mvn-settings: 'aaf-authz-settings'
    files: '**'
    archive-artifacts: ''


Notice that in this example,'{project-name}-{stream}-docker-java-shell-daily' will get translated to 'aaf-authz-arm64-master-docker-java-shell-daily' which is what this change is trying to accomplish.

Patch-set: 1
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1