Add CSIT ARM64 templates
Add specific ARM64 CSIT job templates to avoid issues with
running the corresponding testplans. The plans are hardcoded
to be found under the {project-name} directory. This variable
for ARM is defined as {project-name}-arm64 which makes the dir
location invalid.

ARM64 needs to have its own templates in this case to avoid
modifying the current template to accept testcase-plans as a
parameter and having to define it in every exisiting CSIT job.

Change-Id: Ia7a6d63646800675ae7d2cd165a10c315d623903
Issue-ID: CIMAN-217
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>