Update patch set 2
Patch Set 2:

Yea the purpose of this job was to run pytest. Doing this generates two CSV files, then the last part of the Jenkinks job (that I wasn't able to do in sandbox) is to upload the two CSV files to nexus as raw artifacts.
Is there a way in ONAP today to upload raw artifacts? If so I could redo this to the proper format.

Patch-set: 2
Label: Code-Review=0