Update patch set 2
Patch Set 2:

We don't need to change the current templates... 
In this case, the project and project names should still match the location of the .git folder, which is under dcaegen2/services. 
The only different thing needed to be done is to make sure the location of the pom.xml is defined to the correct subcomponent locations. 

For example, take a look at how other projects do this like:
vfc/vfc-nfvo-wfengine-java.yaml

Just search for "subproject" in ci-management/jjb and you will see several examples

Please test this in the sandbox,, this would not have passed as it is trying to git clone of 'dcaegen2/services/components/datalake-handler'

Patch-set: 2