Update patch set 2
Patch Set 2:

That is the settings file that contain the credentials for both Nexus2 and Nexus3 access.. you need to define it whenever you are using the release-version and docker jobs. 

The name of the settings file should match the name of the repo. In your case, dcaegen2-services-settings

Patch-set: 2