Update patch set 5
Patch Set 5:

Matthieu, Andy, Eric: Can we please approve this change?
This is the job that will replace the current daily release jobs that externalapi is using. 
Since this project is already in oparent 2.0.0 I think this job will work but we can only tell by running it in prod. 

Thanks!!

Patch-set: 5