Update patch set 5
Patch Set 5:

Eric, Currently this jobs is failing since it seems that the latest development in this project is happening on a released artifact:
https://github.com/onap/externalapi-nbi/blob/master/pom.xml#L16

Development must happen on snapshot images and we should let the maven-stage job pick the pom.xml and produce a release artifact every day. 

In general, the gerrit-maven-stage job is promoting an ideal maven workflow for development where the teams need to work on top of snapshots. 

Once this version is changed to a snapshot one, the maven-stage job should work.

Let me know if you have any questions

Patch-set: 5