Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

While this looks good. The namespacing is all wrong. If you submit the template to global-jjb then the namespacing would be correct.

We need to keep things sanely namespaced so that if we find a particular macro / template that is useful to more than a one project we can lift it up to global-jjb without conflicts.

Patch-set: 1
Label: Code-Review=-1