Add multiplatform jobs for policy docker images
Add three global templates needed to build multiarch images:
   - {project-name}-docker-manifest-{stream}: creates a docker
     manifest list with images for multiple architectures by calling
     docker_push_manifest.sh

   - {project-name}-multiarch-docker-stage-{stream}: multijob that
     uses the global-jjb template
     {project-name}-maven-docker-stage-{stream} to produce staging
     images for each supported arch and a manifest list for them

   - {project-name}-multiarch-{stream}-merge-java: multijob that uses
     the existing {project-name}-{stream}-merge-java template to
     produce snapshot images for each suppported arch and a manifest
     list for them

The policy/docker project jobs are modified to use the new multiarch
templates.
All images are now pushed to DockerHub instead of Nexus (as Nexus lacks
support for docker manifest list).

Cancel the triggers for child jobs, as they will be triggered by the
parent multiarch job.

For the jobs that build images, use registry registry-1.docker.io
as a workaround for [1].

Enable the multiarch jobs only for the master branch for now.

[1] https://github.com/fabric8io/docker-maven-plugin/issues/838

Issue-ID: CIMAN-217
Change-Id: Ibd00d24bb70ef51f761005016b7baea4ef66a0a5
Signed-off-by: Paul Vaduva <Paul.Vaduva@enea.com>
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>