Update patch set 1
Patch Set 1: Code-Review+1

This is needed to repair SDNC dockers, which are currently using staged versions of artifacts.  Problem with that is that the staging meta-repo is broken (points at OLDEST instead of NEWEST version), and indications are that it WILL EVER NOT BE FIXED.  So, we must use snapshot dependencies - which work - to create our staging dockers.

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1