Update patch set 20
Patch Set 20: Code-Review-1

Before this one goes in, we need to get the El Alto branch job working. This will need to be re-based to get those changes in. Then create a copy of the "current" set of jobs and remove the master branch from it. Then add in the new multiarch jobs. I can help patch if folks don't understand.

Patch-set: 20
Label: Code-Review=-1