Update patch set 1
Patch Set 1:

just start with aaf, but will do for all the components

The idea is to disable all the CSIT that are neither maintained nor up to date and re-enable them when it makes sense.

Note: it would be very quick to do it from jenkins GUI. I raised a ticket but contributors are not allowed to have jenkins priviledge in ONAP (I had some rights in OPNFV to create view and enable/disable jobs througjh the GUI)

So I will have to patch the different jjbs.

Patch-set: 1
Label: Code-Review=0