Update patch set 1
Patch Set 1:

There seem to be multiple problems here.

The job is malformed because the first mvn goal is missing, I submitted a patch for that: https://gerrit.linuxfoundation.org/infra/c/releng/global-jjb/+/61612

But before that is attempted, the tox run fails with an install error.  I have asked Tommy for help.

Patch-set: 1