Update patch set 1
Patch Set 1:

no this whole effort was in attempt to get away from that. The unit tests now try to run in both. 

tox runs in 37 successfully during the verify jobs. The verify jobs use the python jjb which has 37 fine. The problem is there is some difference between the python jjb and the tox sonar job. we are investigating. 

but backtracking to 36 is definitely not the way.

Patch-set: 1