Update patch set 1
Patch Set 1:

we may have found the issue; or at least we were able to get a similar repo working in the sandbox that had the same error.

there needs to be a patch to `tox-install.sh` in the LF; tox-install currently installs tox using `python`, which points to python2, but then we are running tox against a py3 repo.

we made a ticket: https://jira.linuxfoundation.org/browse/RELENG-2361

we are going to try the same step in the sandbox for the CBS and report back

Patch-set: 1