Update patch set 1
Patch Set 1:

we have answered this question.

the current verify job for the CBS does not actually work.

however, BOTH sonar and the verify job work when we make the change to tox_install.sh. As do ORAN projects.

So, this ticket is the key to all of the issues: https://jenkins.onap.org/sandbox/job/dcaegen2-platform-configbinding-tox-sonar/8/console

Patch-set: 1