Update patch set 2
Patch Set 2:

I just want to disable the job. It is up to the project to decide to delete it or not. They may fix and update the job. Until it is done, the jobs must be disabled because it is misleading the community (short pres planned during next PTL meeting)

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0