Update patch set 2
Patch Set 2:

Yes, so this maven-stage process enforces the maven workflow more strictly. 
This means that you will need to reflect your development towards a snapshot artifact instead of a release version. This job will take your snapshot development and produce a release. 
Basically, your pom has a version:
<version>2.6.0</version>
But, you need to make sure this is a SNAPSHOT version to be able to produce a release from.

Patch-set: 2