Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> hey LF, could you please address this PR?

I actually have a meta question on this. I totally understand that the LF is the owner and sole gatekeeper of the global JJB, which spans across many projects. But why is it the case that project personnel never commit to project specific ci-management repos? In the case here where we are just advancing a pointer; what is the concern on the LF side around project-specific pointer updates? Not trying to sound negative at all, I am purely asking just so that I understand.

Patch-set: 2