Update patch set 2
Patch Set 2: Code-Review+2

In this case in particular, many fixes/changes are happening in self releases. We want to make sure LF carefully upgrades global jjb to prevent 1/2 fixes going through. 
In this case this change should be good to go as the current self release changes are tested and safe.

Patch-set: 2
Label: Code-Review=+2