Update patch set 3
Patch Set 3:

We heard from LF on Friday that ONAP has a PyPI staging repository at https://nexus3.onap.org/repository/PyPi.staging/. We also learned that build minions are configured with a .pypirc file for staging and pypi. Still trying to pin down exactly what's in there for pypi but we can test three of the four jobs while we wait for clarification namely verify, merge and release-verify.

Patch-set: 3