Split PyPI templates for CBS docker client
This module is in a subdirectory, so must specify a gerrit trigger
file path to test this module and package it into a distribution for
upload to PyPI.  But that parameter breaks the release verify and
merge jobs. Work around the problem by using two project defs.
Also configure the non-release merge job with repo key pypy-test.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Issue-ID: DCAEGEN2-1729
Change-Id: I5ab67a95d7e3f40cd2241ef69ea22bf0f9ce9b7f