Revert policy/api multiarch change
The policy jobs are failing because the policy/api is not
being generated and the old SNAPSHOT disappeared from nexus.
Until the docker release strategy has decisions on it I need
to revert back to this in order to generate api SNAPSHOTS and
for our CSIT to succeed.

We can put this back in place when the docker release strategy
is decided upon.

Issue-ID: CIMAN-217
Change-Id: I5815082e066a21fc2eeb79023a5ef93b2b5f9c19
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>