Update patch set 1
Patch Set 1:

Sonar replied to me today saying "When I check the analysis on SonarCloud it looks like target/ui-react/src directory is not indexed by SQ at all. That could explain the difference in coverage, you should understand why this directory is not included anymore." Do you know if maybe this needs to be defined in the pom.xml jacoco plugin?

Patch-set: 1