Update patch set 4
Patch Set 4:

Thanks for reviewing Ittay,Since we do not have control of the plugins in SonarCloud anymore, I had reached out to Sonar team for help since clamp was facing a similar issue. 
They replied with something like this for them ""When I check the analysis on SonarCloud it looks like target/ui-react/src directory is not indexed by SQ at all. That could explain the difference in coverage, you should understand why this directory is not included anymore." "

I am still trying to understand if we need to explicitly point the reports somehow but i am still investigating. 

You know anything about it?

Patch-set: 4