Update patch set 10
Patch Set 10:

(1 comment)

My earlier comments have all been fixed. When I first reviewed this change I want to encourage you to use lf-activate-venv() to create local venvs but it was not released yet.
I still want to make the case for using it. If this change is 'ready to go' I won't press for that now.
The -l option to bash with 'user' venvs really is sketchy.

Patch-set: 10