Update patch set 4
Patch Set 4:

Basically the view can get defined in any file. 
It gets configured to grep for any Jenkins job like "dcaegen2-platform-mod" and list it in the view. If a job with name "*project-name: dcaegen2-platform-mod*" gets created anywhere else, it will still pick it and put it in the view

Patch-set: 4