Update patch set 1
Patch Set 1:

Jess - I tried updating CCSDK and SDNC to use the global-jjb verify and merge jobs, and the verify build is failing because server-id is unset.  I remember having the same problem when I tried this with ccsdk/parent and I believe the change was in the settings file.  Can you please help?

Patch-set: 1
Label: Code-Review=0