Update patch set 1
Patch Set 1:

Thanks Jessica. 

I should add that if there is code or other packages that depend on python3.6 in these ubuntu images we need to add a new source or build locally in the image for python3.6. This patch fixes the build issues caused by absence, but doesn't address adding python3.6 dependency if other software needs it.

Patch-set: 1