Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> We still need the policy/docker jobs to be fixed before we move onto this repo. 
> 
> Remove the generation of the release tags
This has been done with https://gerrit.onap.org/r/c/policy/docker/+/97909
> Delete the images created in docker hub
This needs to go through a request from LF. Do you want all the current images to be removed (as a fresh start) or just the ones with the release tag?
> Go through the official ONAP release process to release image to docker hub
This needs to be done by the policy committers.

Patch-set: 1