Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > We still need the policy/docker jobs to be fixed before we move onto this repo. 
> > 
> > Remove the generation of the release tags
> This has been done with https://gerrit.onap.org/r/c/policy/docker/+/97909

Hmm - ok but I think we are missing a couple. Notably these tags with 2.0.0-arm64 and 2.0.0-amd64 are still being generated:

https://hub.docker.com/layers/onap/policy-base-alpine/2.0.0-arm64/images/sha256-998063417c3979233266f862e61c66aa036d07b4369144db6d99093e55b405df




> > Delete the images created in docker hub
> This needs to go through a request from LF. Do you want all the current images to be removed (as a fresh start) or just the ones with the release tag?

Yes please, let's remove all of them once we fix the tags mentioned above.

> > Go through the official ONAP release process to release image to docker hub
> This needs to be done by the policy committers.

Confirm - once the first 2 steps are done I will go through a release. And then hopefully we are on our way with all this.

Patch-set: 1