Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > We still need the policy/docker jobs to be fixed before we move onto this repo. 
> > > 
> > > Remove the generation of the release tags
> > This has been done with https://gerrit.onap.org/r/c/policy/docker/+/97909
> 
> Hmm - ok but I think we are missing a couple. Notably these tags with 2.0.0-arm64 and 2.0.0-amd64 are still being generated:
> 
> https://hub.docker.com/layers/onap/policy-base-alpine/2.0.0-arm64/images/sha256-998063417c3979233266f862e61c66aa036d07b4369144db6d99093e55b405df
> 
> 
> 
> 
> > > Delete the images created in docker hub
> > This needs to go through a request from LF. Do you want all the current images to be removed (as a fresh start) or just the ones with the release tag?
> 
> Yes please, let's remove all of them once we fix the tags mentioned above.
> 
> > > Go through the official ONAP release process to release image to docker hub
> > This needs to be done by the policy committers.
> 
> Confirm - once the first 2 steps are done I will go through a release. And then hopefully we are on our way with all this.

Removing 2.0.0-arm64 and 2.0.0-amd64 will be trickier to do, given that they are generated with maven (as opposed to the 2.0.0 release tag that will be generated with the release job). None of the images with the arch suffixes are used in any script, nor should they be used in general. Only the manifest lists (the meta-images) are and should be used, so my opinion is that these can be left as is.

Patch-set: 1
Reviewer: Gerrit User 5109 <5109@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0