Update patch set 6
Patch Set 6:

Chris, I tried adding a second stream/branch, in my local environment, and looked at what changed in the results of running jenkins-jobs test, to make sure it makes sense.  You get docker-merge and docker-verify jobs for the additional streams with the correct branch, which is what is expected.  The other 5 jobs shouldn't be duplicated, and they aren't.

Patch-set: 6