Increase thread stack size for Graal in apex-pdp
In the apex-pdp javascript tasks are executed using Graal scripting
support in Java 11. The javascript scripts executed in Java threads. The
default stack size of 1024k is too small and results in stack overflow
exceptions for any Javascript script that is of a reasonable size.

This review increases the stack size for the JVM for apex-pdp builds.

Issue-ID: POLICY-2106
Change-Id: Iea7512939dcd8c1d68d8c37ebd4af63b89ffaa61
Signed-off-by: liamfallon <liam.fallon@est.tech>