Add frankfurt branch jobs for Policy
parent, common, models, api, pap, drools-pdp, drools-applications,
apex-pdp, distribution will be branched.

engine will be branched, but as it will be deprecated in Guilin the
master branch will no longer be supported.

We will branch gui, but not run any jobs on it as it was not released.
Community can clone the branch and compile locally to test the POC in
the state it was at the end of Frankfurt.

CSIT is still pending, need integration/csit to be branched first.

Issue-ID: POLICY-1514
Change-Id: I900e4a7f59528abd93531534d3c3703db3f9b3bb
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>