Update patch set 1
Patch Set 1:

I am not sure what the previous trigger specifies. Can we add merge job as the trigger, so that the CSIT will run everytime when a new code is merged into the repo.

I will change the image versions to "latest", so that when it runs, it will pick up the latest built image.

Patch-set: 1
CC: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>