Update patch set 1
Patch Set 1:

> why it is not possible to rerun CI job from Jenkins (no rerun button in Jenkins menu for dedicated build)? I know that Jenkins has such option, but in ONAP Jenkins this option is turned off. Why?

This is an excellent question that I've wondered myself too. I assume this has been disabled by LF to reduce the load on Jenkins, but I think this is a bit misguided way of doing it (especially when we are running dozens of scheduled CSIT jobs every day that are testing obsolete images)
 
> But I don't want to run CSITs manually :( 

You really, really should (regardless of what's provided by Jenkins). This will be one of my main themes in Honolulu, because it becomes much more important if/when CSITs actually can start blocking merges. 

> I will mark this MR as +1, because current Csit Jobs CI are badly implemented, but for me it will be better to have trigger action after vnfsdk-refrepo-maven-docker-stage-master build.

Note though that you seem to have survived quite well without it until now :)

Patch-set: 1